refactor and clean up
